Uploaded random waypoint generator
This upload contains a file that will generate a series of valid 4D dummy pose to simulate as a waypoint to be fed into emulated_uav/waypoints and emulated_uav/waypointsList such that egh450_path_planning can turn it into a path and publish it to emulated_uav/path. These ROS messages can then be visualised with rviz and shown on screen.